separate getinfo implementation
